Get uodated APIrequests

+++ src/APIRequests.js
@@ -1,11 +1,23 @@
 import axios from 'axios';
 //import React from 'react';
 
-export const SERVER_URL = process.env.REACT_APP_SERVER_URL || "http://localhost:8000"
+export const SERVER_URL = process.env.REACT_APP_SERVER_URL || "http://localhost:8000";
+const TEAM12_URL = "https://true-friends-404.herokuapp.com";
+const TEAM12_CONFIG = {
+    headers: {
+        Authorization: "JWT eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIiwiZXhwIjoxNjc4MzkxNzEzLCJpYXQiOjE2Njk3NTE3MTMsImp0aSI6ImFiNWQ3YWFkZDkzMzQ0ZGFhMTU4YzA4OTA0YmRkOWI5IiwidXNlcl9lbWFpbCI6InRlYW0xM0BtYWlsLmNvbSJ9.-hrs2GQQ8N2e3dgEBUTmvDUoyx_5ZaRqbt6ArIT_bJk",
+        "Content-Type": "application/json"
+    }
+}
+const TEAM19_URL = "https://social-distribution-404.herokuapp.com";
+const TEAM19_CONFIG = {
+    auth: {username: 'Admin13', password: 'password'}, 
+    headers: {"Content-Type": "application/json",}
+};
 
 if (localStorage.getItem("token")) {
     axios.defaults.headers.common = {'Authorization': `Bearer ${JSON.parse(localStorage.getItem("token")).jwt}`}
-    console.log(localStorage.getItem("token"));
+    console.log(JSON.parse(localStorage.getItem("token")).jwt);
 }
 
 export const getPost  = async (authorId, postId) => {
@@ -28,6 +40,7 @@ export const getInbox = async(authorId, page, size) => {
 
 export const sendPublicInbox = async(authorId, postId) => {
     const path = SERVER_URL + `/inbox/public/${authorId}/${postId}`;
+    console.log(path);
     const response = await axios.post(path);
     return response.data;
 };
@@ -39,7 +52,7 @@ export const sendFriendInbox = async(authorId, postId) => {
 };
 
 export const createPostLike = async(likerId, postId) => {
-    const path = SERVER_URL + `/authors/1/posts/${postId}/likes/${likerId}`;
+    let path = SERVER_URL + `/authors/1/posts/${postId}/likes/${likerId}`;
     const response = await axios.post(path);
 
     let liker = await getAuthor(likerId);
@@ -115,6 +128,11 @@ export const deletePostLike = async(likerId, postId) => {
 export const deleteCommentLike = async(likerId, commentId) => {
     const path = SERVER_URL + `/authors/1/posts/1/comments/${commentId}/likes/${likerId}`;
     const response = await axios.delete(path);
+
+    // TEAM 12 NOT IMPLEMENTED
+
+    // TEAM 19 not required
+
     return response.data;
 };
 
@@ -143,13 +161,36 @@ export const getCommentLikes = async(commentId) => {
 };
 
 export const getAuthor = async(authorId) => {
-    const path = SERVER_URL + `/authors/${authorId}`;
-    const response = await axios.get(path);
-    return response.data;
+    let path = SERVER_URL + `/authors/${authorId}`;
+    let response = await axios.get(path);
+    if (response.data !== "Author does not exist") { return response.data; }
+
+    // TEAM 12
+    path = TEAM12_URL + `/authors/${authorId}/`;
+    response = await axios.get(path, TEAM12_CONFIG);
+    if (!response.data.detail) { 
+        response.data.type = "author";
+        response.data.url = `${response.data.host}/authors/${authorId}/`;
+        response.data.accepted = true;
+        response.data.profileImage = response.data.profile_image;
+        response.data.host = TEAM12_URL;
+        return response.data 
+    }
+
+    // TEAM 19
+    path = TEAM19_URL + `/authors/${authorId}`;
+    response = await axios.get(path, TEAM19_CONFIG);
+    if (response.data !== "Author matching query does not exist.") { 
+        response.data.accepted = true;
+        response.data.host = TEAM19_URL;
+        return response.data; 
+    }
+
+    return "Author does not exist";
 };
 
 export const createPost = async (authorId, data) => {
-    const path = SERVER_URL + `/authors/${authorId}/posts`;
+    let path = SERVER_URL + `/authors/${authorId}/posts`;
     const response = await axios.put(`${path}`,data);
 
     let author = await getAuthor(authorId);
@@ -221,7 +262,7 @@ export const createPost = async (authorId, data) => {
 }
 
 export const createComment = async (authorId, postId, data) => {
-    const path = SERVER_URL + `/authors/${authorId}/posts/${postId}/comments`;
+    let path = SERVER_URL + `/authors/${authorId}/posts/${postId}/comments`;
     const response = await axios.post(`${path}`,data);
 
     let author = await getAuthor(authorId);
@@ -264,8 +305,29 @@ export const searchForAuthors = async (query, page, size) => {
     if (size !== null) {
         path += `&size=${size}`;
     }
-
     const response = await axios.get(path);
+
+    // TEAM 12
+    path = TEAM12_URL + `/authors/`;
+    let team12 = (await axios.get(path, TEAM12_CONFIG)).data;
+    for (let author of team12) {
+        author.type = "author";
+        author.displayName = author.username;
+        author.github = null;
+        author.accepted = author.is_active;
+        author.profileImage = "https://t3.ftcdn.net/jpg/05/16/27/58/360_F_516275801_f3Fsp17x6HQK0xQgDQEELoTuERO4SsWV.jpg";
+        response.data.authorsPage.push(author);
+    }
+
+    // TEAM 19
+    path = TEAM19_URL + `/authors`;
+    let team19 = (await axios.get(path, TEAM19_CONFIG)).data;
+    for (let author of team19.items) {
+        author.accepted = true;
+        author.id = author.id.split('/authors/')[1];
+        response.data.authorsPage.push(author);
+    }
+
     return response.data;
 }
 
@@ -276,7 +338,6 @@ export const modifyAuthor = async (authorId, newGithub, newProfileImage) => {
         "github": newGithub,
         "profileImage": newProfileImage
     }
-    console.log(data);
     //console.log(GetAuthDetails());
     axios.post(path, data);//, {auth: GetAuthDetails()});
 
@@ -293,8 +354,6 @@ export const checkFollowStatus = async (authorId, foreignAuthorId) => {
     const followRequestResponse = await axios.get(`${followRequestPath}`);
 
     let followStatus = -1
-    console.log(followResponse.data)
-
     if (followResponse.data.id !== "") {
         followStatus = 2
     }
@@ -348,9 +407,21 @@ export const requestToFollow = async (authorId, foreignAuthorId) => {
 }
 
 export const removeFollower = async (authorId, foreignAuthorId) => {
-    const path = SERVER_URL + `/authors/${foreignAuthorId}/followers/${authorId}`
+
+    let foreignAuthor = await getAuthor(foreignAuthorId);
+    if (foreignAuthor === "Author does not exist") { return foreignAuthor; }
+    
+    let path = SERVER_URL + `/authors/${foreignAuthorId}/followers/${authorId}`;
     const response = await axios.delete(`${path}`);
-    return response.status
+
+    if (foreignAuthor.host === TEAM12_URL) {
+        // TEAM 12 NOT IMPLEMENTED
+    } else if (foreignAuthor.host === TEAM19_URL) {
+        // TEAM 19 has no logic required
+        path = TEAM19_URL + `/authors/${foreignAuthorId}/followers/${authorId}`
+        axios.delete(path, TEAM19_CONFIG);
+    }
+    return response.data;
 }
 
 export const getFollowRequests = async (authorId) => {
@@ -377,7 +448,7 @@ export const addFollower = async (authorId, foreignAuthorId) => {
     return response.data;
 }
 
-export const removeFollowRequest = async (authorId, foreignAuthorId) => {
+export const removeFollowRequest = async (authorId, foreignAuthorId) => {   
     const path = SERVER_URL + `/authors/${authorId}/followRequest/${foreignAuthorId}`
     const response = await axios.delete(`${path}`);
  
@@ -416,13 +487,19 @@ export const authUser = async (data) => {
 }
 
 export const deletePost = async(authorId, postId) => {
-    const path = SERVER_URL + `/authors/${authorId}/posts/${postId}`
+    let path = SERVER_URL + `/authors/${authorId}/posts/${postId}`
     const response = await axios.delete(path);
+
+    // TEAM 12
+    path = TEAM12_URL + `/posts/${postId}/`;
+    axios.delete(path, TEAM12_CONFIG);
+
+    // TEAM 19 has no logic required for this
     return response.data;
 };
 
 export const modifyPost = async (authorId, postId, data) => {
-    const path = SERVER_URL + `/authors/${authorId}/posts/${postId}`
+    let path = SERVER_URL + `/authors/${authorId}/posts/${postId}`;
     axios.post(path, data);
 
     // TEAM 12
@@ -462,8 +539,10 @@ export const getImage = async (imageUrl) => {
     return response.data;
 }
 
-export const getGithubEvents = async (githubUsername) => {
+export const getGithubEvents = async (github) => {
+    let githubUsername = github.split(".com/")[1];
+    if (!githubUsername) { githubUsername = github; }
     const path = `https://api.github.com/users/${githubUsername}/events`;
-    let response = await axios.get(path);
+    let response = await axios.get(path, {headers: {Authorization: ""}});
     return response.data;
 }
